Create and update an "active" floreant directory on the Desktop
The 'live' version of the build is now copied to

~/Desktop/active_floreantpos

for anyone who might want a floreant installation they can just start
running (skipping the step of installing the .zip).  Note that this
directory will have all files overwritten, including the logo.png
file.  I've submitted a patch to the devs to hopefully address this.